FollowLineAlternative
An alternative to line following. The idea is to move perfectly straight until the line sensor moves off the line. This isn't in the code right now, but then it would stop, turn left until the other side of the line sensor falls off, and use the gyro sensor to bisect the angles and end up straight on the line again. The idea is to update theta whenever it has to turn so it moves straight relative to the angle it thinks is best to move.